refactor: reorganize event package and consumer logic
- Rename package from 'urlshort' to 'event' in internal/event files
- Move StartConsumers function from deps.go to consumer.go
- Add missing os import in consumer.go
- Remove unused imports from deps.go
