+++ api/internal/event/http.go
@@ -1,4 +1,4 @@
-package urlshort
+package event
 
 import (
 	"net/http"
