+++ api/internal/event/service.go
@@ -1,4 +1,4 @@
-package urlshort
+package event
 
 import (
 	"context"
