+++ pkg/deps/consumer.go
@@ -2,6 +2,7 @@ package deps
 
 import (
 	"context"
+	"os"
 
 	amqp "github.com/rabbitmq/amqp091-go"
 	"github.com/vit0rr/publish-and-consume/config"
@@ -27,6 +28,23 @@ func NewConsumer(deps *Deps, amqpConn *amqp.Connection, db *mongo.Database, conf
 		RabbitMQ: amqpConn,
 		Mongo:    db,
 	}, nil
+}
+
+func StartConsumers(ctx context.Context, cancel context.CancelFunc, cfg config.Config, mongo *mongo.Client, amqpConn *amqp.Connection, dependencies *Deps) error {
+	eventsConsumer, err := NewConsumer(dependencies, amqpConn, mongo.Database("db_events"), cfg, ctx)
+	if err != nil {
+		log.Error(ctx, "failed to create events consumer", log.ErrAttr(err))
+		os.Exit(1)
+	}
+
+	go func() {
+		if err := eventsConsumer.Start(ctx); err != nil {
+			log.Error(ctx, "consumer error", log.ErrAttr(err))
+			cancel()
+		}
+	}()
+
+	return nil
 }
 
 func (c *Consumer) Start(ctx context.Context) error {
